Handle Service Requests more gracefully
The existing code didn't format service requests particularly well since those tickets
have more information that requires additional API calls to retrieve. This commit adds
API calls and support to get those additional fields, and then displays them if the
ticket type is an SR.

A TODO from this is that TicketStatus now has some custom fields as part of it, so
that needs to either be configurable or built into the end user's formatting code.